Fix for the loading issue
it seems that when duplicating the growing behaviour I forgot to chenge the type returned. That caused the load to load both lifecycle and dying behaviours as growing behaviours and that caused the saved memory to be out of order since the number a types of values loaded for each behaviour are different. Upon testing I couold see no issues at the moment so the issue seems to be fixed.