Created a scene for the prototype
Created a scene with four worlds and a goal worlds, although it is possible to traverse the level the code is not there to detect the plannets. The tutorials were really usefull for this one. Next time I will create the WorldScript that detects when a player falls on the surface and displays a message with the problem or the win message.
A variant of the sphereMovement would also be good to make randomly moving ai with a simple trigger that quits the game or something.
I also need to add a way to exit the game from within.