Walls now connect diagonally
Extended the Game Tile to hold references to all surroundingtiles, including diagonally. Walls now adapt to the presence of other walls arround them but not towers. We might keep towers as separate entities. I am thinking maybe enemies can move through them, but if they do they deal damage to it as if it was a destination