Resource generation and bridges are in
Implemented the generation of resources in a dumb way, but it works and it is configurable to an extent.
I also implemented bridges, but it is still possible to get a river without a bridge if it is in a diagonal. It is not too important for the prototype that the map generation have no flaws.

Next i will implement resource counter, resource gathering and replcaing resources with constructions in the map. Not sure if I will work on building costs for the prototype. maybe after the resource gathering is working.