Resource generation is in
*This change cause a peak on map generation time. Will have to investigate, but eficiency is not extremely important for the prototype.*
Resources are generated at a constant speed of 0.1/resourceTile of a single type. Gold cannot be generated, we would make a trade system that allows the player to trade resources for gol with the empire to the north.
Destinations now adapt to tile changes by recalculating their resource generation. Multiple destinations can gather the same resource tile.